requirements.txt
python-dotenv==1.0.1
requests==2.32.3
base58==2.1.1
